ENGR00125796 MX5x: Boot stop at "Starting kernel ..."
System sometime stop at "Starting kernel ...". After connect with
ICE, we found that system enter abort mode. The abort sometimes is
data abort and sometimes prefetch abort.This issue may caused by that
we config the PLL1 while CPU clock is runing on it.In fact, we don't
need do PLL1 enable operation if it already run up and which will
bring side-effect for the impact of cpu clock.

Signed-off-by:Jason Liu <r64343@freescale.com>