ENGR00134179 mx50 rdp: set VSD regulator always on
This is required by SD2 card detection function.

The root cause is that the power for SD2 card detection pull up
is the same as the SD2 I/O power, they are all supplied by VSD
regulator. If we disable VSD dynamically, the card will never be
able to be detected again.

Signed-off-by: Aisheng.Dong <b29396@freescale.com>