ENGR00144146 asrc: add help info in Kconfig for compiling as module
The ASOC platform driver(built in kernel) calls functions
defined in ASRC, So the kernel may fail to compile when
config ASRC as module.

The solution is also compiling ASOC platform driver as module.
It can not be done by memuconfig system automatically
since ASOC platform does not depends on ASRC driver.
So we have to add a help info in Kconfig to tell user about
this limitation.

Signed-off-by: Dong Aisheng <b29396@freescale.com>