ENGR00144104-2 AHCI: Some usb related clocks are always on even without USB disk
In order to decrease the pwr consumption, release the CLK
resources such as usb_phy1_clk, when there is no SATA device
adaptored into the AHCI SATA port. The HOTPLUG feature can't
be enabled in this situation.

Modify the AHCI confiuration to control the configurations of
AHCI_SAVE_PWR_WITHOUT_HOTPLUG macro used in AHCI driver.

Signed-off-by: Richard Zhu <r65037@freescale.com>