ENGR00151896 - EPDC fb: Fix bug preventing suspend when powerdown disabled
- With FB_POWERDOWN_DISABLE is set as the powerdown delay (disabling the EPDC
driver from powering down), the EPDC driver was having problems going into
suspend.  Any request to blank the EPDC driver would result in a timeout
and an error status.  And subsequent attempts to send updates to the driver
would fail, as the driver's power state was incorrect.  This patch prompts
the EPDC driver to be disabled when the blanking state is changed to
FB_BLANK_POWERDOWN and FB_POWERDOWN_DISABLE is set.  Resultingly, the correct
driver power state is maintained and suspend/resume works correctly.

Signed-off-by: Danny Nold <dannynold@freescale.com>