ENGR00162711 DA9053: Add dummy write for DA9053 I2C register access
DA9053 i2c issue: Rarely the i2c interface of DA9053 hang and it can
not be recovered if not power off totally. The Dialog suggests adding
dummy write for DA9053 I2C register access, in order to decrease the failure
of DA9053 register access and possibility of i2c failure.

Signed-off-by: Wayne Zou <b36644@freescale.com>