ENGR40464722-1 MX28 ALSA: wait until the DMA channel finishs
When we invoke mxs_pcm_close(), the DMA channel may still does
some DMA operations. We should wait for the DMA channel to finish
all the DMA operations. We have observed a gplay hang if we do not do this.

How the hang happen?
  [1] the DMA channel is not closed properly.
  [2] So the DMA is in a mess state, when we play video with gplay
      in the next time. The audio system does not work, which causes
      the video system does not work too, so the gplay hang.

How to close the DMA channel properly:
  [1] wait until the DMA channel finishs all the DMA operations.
  [2] do not reset the DMA channel here, it makes no sense.

Signed-off-by: Huang Shijie <b32955@freescale.com>