ENGR00223169 MX508-Ensure TLBs are primed before ddr enters self-refresh.
Need to ensure that no page table walk occurs in DDR when
the clock to DDR is lowered when system enters WFI. Hence
ensure that the TLB contains the addresses of the registers
that are accessed in the WFI routine.
Also add a dsb instruction to ensure all write buffers are
drained before DDR is in self-refresh.

Signed-off-by: Ranjani Vaidyanathan <ra5478@freescale.com>