ENGR00223524 MX508-Fix the incorrect reset by SRTC
SRTC power inputs are 1.2V coin-cell and 1.2V LDO output (1p2).
We need to switch from coin-cell to 1P2 when the VDD1P2_good
signal goes high. And VDD1p2_good signal is not valid unless
bandgap is up. When there is no bandgap, its possible that the
VDD1P2_good signal goes high when the LDO output is only 0.5V
which causes SRTC to issue a reset.

To ensure that the SRTC does not incorrectly switch to 1P2,
the following steps should be followed:
1. Donot power down the bandgap explicitly in SW.
2. Disable the 1P2 brown out detector. When brown out detector
is disabled, VDD1P2_good signal is 0.
3. Clear the bandgap REF_SELFBIAS_OFF bit just before system
enters suspend in case the bandgap powers down if its supply is
disabled externally.
4. After resume, set the bandgap REF_SELFBIAS_OFF bit. This
should be set 100usec after the bandgap powers up. But by the
time the SW gets control, OSC has stablized which will take 8msec,
so we don't have to delay another 100usecs.
5. Enable the 1P2 brown out detector.

Signed-off-by: Ranjani Vaidyanathan <ra5478@freescale.com>