hdmidongle: defconfig for working host-side USB
The OTG port on the SOC is connected to both the micro B (device port,
also power port) and regular A (host) port. This kernel tries to use the
OTG ID pin to determine whether a host or device is connected; the
hdmidongle does not connect the ID pin to the SOC's OTG ID pin(s).

Instead, we build OTG support as modules. The system will boot in USB
host mode, and can be changed to device mode by loading the relevant
gadget modules.